+++ src/view/conversationEditor/utils/readYaml.tsx
@@ -1,6 +1,6 @@
 import { MarkerType, Node, Edge } from "reactflow";
 import * as yaml from "js-yaml";
-import { arrayAppend, logWarning, stringSplitToArray } from "./commonUtils";
+import { arrayAppend, stringSplitToArray } from "./commonUtils";
 import ConversationYamlModel from "./conversationYamlModel";
 
 export interface YamlReaderOutput {
@@ -10,15 +10,17 @@ export interface YamlReaderOutput {
 
 export function readYaml(
   fileName: string,
-  text: string
+  text: string,
+  translationSelection: string,
 ): YamlReaderOutput | null {
   const data = yaml.load(text) as ConversationYamlModel;
-  return parseYaml(fileName, data);
+  return parseYaml(fileName, data, translationSelection);
 }
 
 export function parseYaml(
   fileName: string,
-  yaml: ConversationYamlModel
+  yaml: ConversationYamlModel,
+  translationSelection: string
 ): YamlReaderOutput | null {
   // Load
   const npcOptions = yaml.NPC_options;
@@ -34,7 +36,7 @@ export function parseYaml(
     id: "startNodeID",
     type: "startNode",
     position: { x: 0, y: 0 },
-    data: { name: "start", text: fileName, text2: yaml.quester, text3: yaml.stop },
+    data: { name: "start", fileName: fileName, yaml: yaml, translationSelection: translationSelection },
   };
   const startNodes: Record<string, Node> = { startNodeID: startNode };
 
@@ -60,10 +62,11 @@ export function parseYaml(
       position: { x: 0, y: 0 },
       data: {
         name: key,
-        text: option.text,
+        option: option,
         events: stringSplitToArray(events),
         conditions: stringSplitToArray(conditions),
         pointers: pointers,
+        translationSelection: translationSelection,
       },
     };
     npcNodes[idKey] = dict;
@@ -91,10 +94,11 @@ export function parseYaml(
       position: { x: 0, y: 0 },
       data: {
         name: key,
-        text: option.text,
+        option: option,
         events: stringSplitToArray(events),
         conditions: stringSplitToArray(conditions),
         pointers: pointers,
+        translationSelection: translationSelection,
       },
     };
     playerNodes[idKey] = dict;